backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["RomanL91 <dev_roman91@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.110.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
